+++ src/app/register/nicovideo/page.tsx
@@ -1,12 +1,12 @@
 import clsx from "clsx";
 
-import { Form2 } from "./Form2";
+import { Form } from "./Form";
 
 export default async function Page() {
   return (
     <main>
       <h1>Register</h1>
-      <Form2 className={clsx(["mt-4"])} />
+      <Form className={clsx(["mt-4"])} />
     </main>
   );
 }
